+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.0
+* Bump Android native library to 1.6.0 
+* Fix Volumedeck background service for Android 14+
+
 ## 1.1.6
 * Improve readme